made the integer to double as default type

+++ Matrix/MatrixOverloads.cpp
@@ -35,8 +35,9 @@ Matrix& Matrix::operator=(Matrix& matrix){
 	return *this;
 }
 
+
 //scaler multiply matrix * scaler 
-Matrix Matrix::operator*(int scaler){
+Matrix Matrix::operator*(double scaler){
 	Matrix resultMatrix(this->getRows(),this->getColumns());
 	for(int i = 1 ; i <= this->getRows() ; i++){
 		for(int j = 1 ; j <= this->getColumns() ; j++){
@@ -47,12 +48,12 @@ Matrix Matrix::operator*(int scaler){
 }
 
 //friend function to implement scaler * matrix
-Matrix operator*(int scaler , Matrix& matrix){
+Matrix operator*(double scaler , Matrix& matrix){
 	return (matrix * scaler);
 }
 
 //scaler divide matrix * scaler 
-Matrix Matrix::operator/(int scaler){
+Matrix Matrix::operator/(double scaler){
 	Matrix resultMatrix(this->getRows(),this->getColumns());
 	for(int i = 1 ; i <= this->getRows() ; i++){
 		for(int j = 1 ; j <= this->getColumns() ; j++){
@@ -63,7 +64,7 @@ Matrix Matrix::operator/(int scaler){
 }
 
 //friend function to implement scaler / matrix
-Matrix operator/(int scaler , Matrix& matrix){
+Matrix operator/(double scaler , Matrix& matrix){
 	return (matrix / scaler);
 }
